project/requirements.txt
Django==3.0.4
psycopg2-binary==2.8.4
gunicorn==20.0.4
django-modeltranslation==0.15
martor==1.4.9
django-autoslug==1.9.7
